Move more components into specific folders
