+++ src/components/Filters/index.js
@@ -1,0 +1,2 @@
+import Filters from './Filters';
+export default Filters;
