+++ src/components/View/index.js
@@ -1,0 +1,2 @@
+import View from './View';
+export default View;
